% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/creditrISDA-package.R
\docType{package}
\name{creditrISDA}
\alias{creditrISDA}
\alias{creditrISDA-package}
\title{The creditrISDA package.}
\description{
\code{creditrISDA} package contains the C code of ISDA Standard Model, which is developed
for evaluating credit default swaps (CDS). After Big Bang Protocol came out in 2009
after the financial crysis, CDS trading is standardized and ISDA Standard Model
becomes a widely used approach to price CDS and measure risk exposures of CDS.
}
\details{
\code{creditrISDA} package contains the C code of the original open-source ISDA Model,
as well as the R wrapper to call ISDA Model with R functions, using \code{Rcpp}. This
package is intended to be a helper package for R package \code{creditr}, which is
available on CRAN.
}

