% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upfront.to.spread.R
\name{upfront.to.spread}
\alias{upfront.to.spread}
\title{Calculate Spread with a Given Upfront}
\usage{
upfront.to.spread(x, currency.var = "currency", date.var = "date",
  coupon.var = "coupon", tenor.var = "tenor", maturity.var = "maturity",
  recovery.var = "recovery", upfront.var = "upfront",
  points.var = "ptsUpfront", isPriceClean = FALSE, notional = 1e+07,
  payAccruedAtStart = FALSE, payAccruedOnDefault = TRUE)
}
\arguments{
\item{upfront.var}{is the character name of upfront column}

\item{points.var}{character name of points Upfront column}

\item{isPriceClean}{a boolean variable indicating whether the upfront is clean or dirty}

\item{notional}{numeric variable indicating the notional value of the CDS contract}

\item{payAccruedAtStart}{whether pay at start date the accrual amount}

\item{payAccruedOnDefault}{whether pay in default scenario the accrual amount}
}
\value{
a numeric indicating the spread.
}
\description{
\code{upfront.to.spread} calculates conventional spread using the upfront
or ptsUpfront values.
}

